Safari-609.1.11.1/LayoutTests/media/media-play-promise-reject-pause-abort-expected.txt


RUN(mediaElement.play().then(failTest).catch(promiseRejected))
RUN(mediaElement.pause())
Promise rejected. OK
EXPECTED (error.name == 'AbortError') OK
END OF TEST
